Move admin controllers into their own directory. Keep existing working for backwards compatibility
